moves sequencer and controls to instruments folder
