+++ src/components/Instruments/Sequencer/Channel.css
