+++ src/components/Instruments/Sequencer/Controls.jsx
