+++ src/components/Instruments/Sequencer/Sequencer.css
